OBSOLETE: BACKPORT: FEDORA: lis3: improve handling of null rate
When obtaining a rate of 0, we would disable the device supposely
because it seems to behave incorectly. It actually only comes from the
fact that the device is off and on lis3dc it's reflected in the rate.
So handle this nicely by just waiting a safe time, and then using the
device as normally.

Bugzilla: 785814
Upstream-status: ??

Signed-off-by: Éric Piel <eric.piel@tremplin-utc.net>

[Kexy: Resolved minor conflict in drivers/misc/lis3lv02d/lis3lv02d.c]
Link: https://src.fedoraproject.org/rpms/kernel/c/3d246278f1c64d7e1d5ab46d7b885a3aafd2cd66
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>