OBSOLETE: FEDORA: lib/cpumask: Make CPUMASK_OFFSTACK usable without debug dependency
When CPUMASK_OFFSTACK was added in 2008, it was dependent upon
DEBUG_PER_CPU_MAPS being enabled, or an architecture could select it.
The debug dependency adds additional overhead that isn't required for
operation of the feature, and we need CPUMASK_OFFSTACK to increase the
NR_CPUS value beyond 512 on x86.  We drop the current dependency and make
sure SMP is set.

Bugzilla: N/A
Upstream-status: Nak'd, supposedly replacement coming to auto-select

Signed-off-by: Josh Boyer <jwboyer@fedoraproject.org>

Link: https://src.fedoraproject.org/rpms/kernel/c/9742e529b878106143a0a68c4734af4328688a99?branch=rawhide
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>