AOSCOS: Input: i8042 - Make i8042_bypass_aux_irq_test as a module parameter
The original i8042_bypass_aux_irq_test is a variable which cannot be
changed without kernel rebuild. Some buggy platform fails to detect the
existance of AUX if i8042_bypass_aux_irq_test isn't set. So make this
variable as a module parameter and set the default value to true.

Signed-off-by: Huacai Chen <chenhc@lemote.com>

Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>