drm: sun4i: set the dotclock when enabling the dotclock
Currently the dotclock is set when setting the mode of TCON. After this
the dotclock may gets changed, e.g. when the downstream encoder try to
change another clock under the same parent.

As the dotclock is saved by CRTC state, set it when the clock is about
to be enabled, and use clk_set_rate_exclusive() to get the exclusivity
at the same time. This will prevent any unexpected changes to the
dotclock.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>