drm: sun4i: set the sun8i-dw-hdmi TMDS clock exclusively when enabling
On H3/H5/A64/R40 the HDMI PHY clock is set after the the mode_set() of
the HDMI encoder is called. As the PHY clock has the same parent with
the TMDS clock (PLL-VIDEO{0,1}), it may mess up TMDS clock.

Set the rate of TMDS clock only when enabling, and add exclusivity to
it. Drop the exclusivity when disabling, to allow the change of other
clocks connected under the PLL.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>