clk: sunxi-ng: Consider limits for recalc callback
Currently, if NM PLL is set to unstable rate before Linux kernel is
loaded, it still can be cosidered as valid rate. This can happen if PLL
rate is multiply of wanted rate.

For example, if PLL is set to 128 MHz (below 192 MHz threshold) and
target rate is 32 MHz, clk framework will happily use that.

To counteract this, return 0 as currently set frequency if recalc_rate
function if actuall PLL frequency is outside of working (stable) range.

Signed-off-by: Jernej Skrabec <jernej.skrabec@siol.net>
[Icenowy: fix a use-before-check error and change returned 1 to 0]
Signed-off-by: Icenowy Zheng <icenowy@aosc.io>