arm64: allwinner: h6: pine-h64: add PIO VCC bank supplies
The Allwinner H6 SoC features tweakable VCC for PC, PD, PG, PL and PM
banks.

This patch adds supplies for PC and PD banks. PG and PM banks are
used for Wi-Fi and should be added when Wi-Fi is added; PL bank is where
PMIC is attached, and currently if a PMIC regulator is added for it
dependency loop will happen.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>