arm64: allwinner: a64: dts: add volumn keys support for PineTab
PineTab has a pair of volumn keys (up and down) which are connected via
a resistor ladder on the LRADC pin of the A64 SoC. Volumn + key is
active at 0.2V and Volumn - is at 0.4V according to calculation.

Add volumn keys support for PineTab by enabling the LRADC and fill key
voltage infomation.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>