BACKPORT: FROMLIST: mips/mm: Add NUMA balancing support
NUMA balancing is available on nearly all architectures, but MIPS lacks
it for a long time. In theory, the current NUMA balancing framework only
need a "PROTNONE" page table bit and some APIs to manipulate it. So, it
is time for us to add MIPS's NUMA balancing support (Only for 64bit now
because NUMA balancing depends on huge page implicitly).

Signed-off-by: Huacai Chen <chenhc@lemote.com>

Link: https://lore.kernel.org/all/1604201638-4001-1-git-send-email-chenhc@lemote.com/
[Kexy: Resolved minor conflicts in arch/mips/Kconfig, and
 arch/mips/include/asm/pgtable.h]
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>