BACKPORT: SURFACE: hid/multitouch: Turn off Type Cover keyboard backlight when suspending
The Type Cover for Microsoft Surface devices supports a special usb
control request to disable or enable the built-in keyboard backlight.
On Windows, this request happens when putting the device into suspend or
resuming it, without it the backlight of the Type Cover will remain
enabled for some time even though the computer is suspended, which looks
weird to the user.

So add support for this special usb control request to hid-multitouch,
which is the driver that's handling the Type Cover.

The reason we have to use a pm_notifier for this instead of the usual
suspend/resume methods is that those won't get called in case the usb
device is already autosuspended.

Also, if the device is autosuspended, we have to briefly autoresume it
in order to send the request. Doing that should be fine, the usb-core
driver does something similar during suspend inside choose_wakeup().

To make sure we don't send that request to every device but only to
devices which support it, add a new quirk
MT_CLS_WIN_8_MS_SURFACE_TYPE_COVER to hid-multitouch. For now this quirk
is only enabled for the usb id of the Surface Pro 2017 Type Cover, which
is where I confirmed that it's working.

Patchset: surface-typecover

Signed-off-by: Jonas Dreßler <verdre@v0yd.nl>
Link: https://github.com/linux-surface/kernel/commit/04516807cb12c7021afbc2126e651ed4f97f6ae7

Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>

[Mingcong Bai: Resolved a minor merge conflict in
 drivers/hid/hid-multitouch.c]

Signed-off-by: Mingcong Bai <jeffbai@aosc.io>