SURFACE: platform/surface: Add driver for Surface Book 1 dGPU switch
Add driver exposing the discrete GPU power-switch of the  Microsoft
Surface Book 1 to user-space.

On the Surface Book 1, the dGPU power is controlled via the Surface
System Aggregator Module (SAM). The specific SAM-over-HID command for
this is exposed via ACPI. This module provides a simple driver exposing
the ACPI call via a sysfs parameter to user-space, so that users can
easily power-on/-off the dGPU.

Patchset: surface-sam-over-hid

Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>
Link: https://github.com/linux-surface/kernel/commit/82a9df58b23b3f6d0486f3731036f4180f216ae3
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>