SURFACE: platform/surface: Add driver for Surface Book 1 dGPU switch
Add driver exposing the discrete GPU power-switch of the  Microsoft
Surface Book 1 to user-space.

On the Surface Book 1, the dGPU power is controlled via the Surface
System Aggregator Module (SAM). The specific SAM-over-HID command for
this is exposed via ACPI. This module provides a simple driver exposing
the ACPI call via a sysfs parameter to user-space, so that users can
easily power-on/-off the dGPU.

Patchset: surface-sam-over-hid

Link: https://github.com/linux-surface/linux-surface/commit/91786c371cc7d1c49ca17679243efd614b22e7ec
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>