SURFACE: ath10k: Add module parameters to override board files
Some Surface devices, specifically the Surface Go and AMD version of the
Surface Laptop 3 (wich both come with QCA6174 WiFi chips), work better
with a different board file, as it seems that the firmeware included
upstream is buggy.

As it is generally not a good idea to randomly overwrite files, let
alone doing so via packages, we add module parameters to override those
file names in the driver. This allows us to package/deploy the override
via a modprobe.d config.

Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>
Patchset: ath10k

Link: https://github.com/linux-surface/linux-surface/blob/26b04e31e452a48e9682e22b81215a73c962b8e9/patches/6.15/0004-ath10k.patch
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>