SURFACE: ACPI: Add quirk for Surface Laptop 4 AMD missing irq 7 override
This patch is the work of Thomas Gleixner <tglx@linutronix.de> and is
copied from:
https://lore.kernel.org/lkml/87lf8ddjqx.ffs@nanos.tec.linutronix.de/

This patch adds a quirk to the ACPI setup to patch in the the irq 7 pin
setup that is missing in the laptops ACPI table.

This patch was used for validation of the issue, and is not a proper
fix, but is probably a better temporary hack than continuing to probe
the Legacy PIC and run with the PIC in an unknown state.

Patchset: amd-gpio

Signed-off-by: Sachi King <nakato@nakato.io>
Link: https://github.com/linux-surface/kernel/commit/b410202aa6f79828419e492937e64bdac379bda0
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>