arm64: dts: apple: Add initial t6000/t6001/t6002 DTs
t6002 is two die version of t6001. Devices on the second die are at a
consistent offset compared to the first die. Use "ranges" properties
to translate the MMIO addresses. To avoid duplication all devices are
separate *.dtsi. C preprocessor macros are used to assign unique node
names and labels.
Devices are split into devices present on die 0, devices present on all
dies and NVMe.
It's unclear if devices on die 1 are fused out or just not used on the
single machine with t6002.

The M1 Ultra SoC consists of two M1 Max dies. The multi die device
nodes are distributed to include files based on whether they are used
on both dies or just the primary die on the Mac Studio M1 Ultra.
The NVMe is a noteable exception since it is only on the second die.
It seems to be a special case since its interrupt is routed as it were
on die 0.

t6000 is a cut-down version of t6001, so the former just includes the
latter and disables the missing bits (This is currently just one PMGR
node and all of its domains.

Signed-off-by: Hector Martin <marcan@marcan.st>
Co-developed-by: Janne Grunau <j@jannau.net>
Signed-off-by: Janne Grunau <j@jannau.net>