SURFACE: USB: quirks: Add USB_QUIRK_DELAY_INIT for Surface Go 3 Type-Cover
The touchpad on the Type-Cover of the Surface Go 3 is sometimes not
being initialized properly. Apply USB_QUIRK_DELAY_INIT to fix this
issue.

More specifically, the device in question is a fairly standard modern
touchpad with pointer and touchpad input modes. During setup, the device
needs to be switched from pointer- to touchpad-mode (which is done in
hid-multitouch) to fully utilize it as intended. Unfortunately, however,
this seems to occasionally fail silently, leaving the device in
pointer-mode. Applying USB_QUIRK_DELAY_INIT seems to fix this.

Link: https://github.com/linux-surface/linux-surface/issues/1059
Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>
Patchset: surface-typecover

Link: https://github.com/linux-surface/linux-surface/blob/26b04e31e452a48e9682e22b81215a73c962b8e9/patches/6.15/0010-surface-typecover.patch
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>