SURFACE: USB: quirks: Add USB_QUIRK_DELAY_INIT for Surface Go 3 Type-Cover
The touchpad on the Type-Cover of the Surface Go 3 is sometimes not
being initialized properly. Apply USB_QUIRK_DELAY_INIT to fix this
issue.

More specifically, the device in question is a fairly standard modern
touchpad with pointer and touchpad input modes. During setup, the device
needs to be switched from pointer- to touchpad-mode (which is done in
hid-multitouch) to fully utilize it as intended. Unfortunately, however,
this seems to occasionally fail silently, leaving the device in
pointer-mode. Applying USB_QUIRK_DELAY_INIT seems to fix this.

Link: https://github.com/linux-surface/linux-surface/issues/1059
Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>
Patchset: surface-typecover

Link: https://github.com/linux-surface/linux-surface/tree/54cbb6c15e5fd226f8ff19af15e7c3dc3eb86f3d
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>