drm/asahi: alloc: Support tagging array allocs
It's hard to tell what a given array buffer is just from the type, so
add support for explicitly adding a u32 tag. This can help us
differentiate between allocs in the debug codepaths or when dumping
memory.

To more easily debug GPU/FW-side overreads, use the alloc tag to fill
the padding instead of using a constant.

Signed-off-by: Asahi Lina <lina@asahilina.net>