iommu: apple-dart: Power on device when handling IRQs
It's possible for an IRQ to fire and the device to be RPM suspended
before we can handle it, which then causes device register accesses to
fail in the IRQ handler.

Since RPM is IRQ-safe for this device, just make sure we power on the
DART in the IRQ handler too.

Signed-off-by: Asahi Lina <lina@asahilina.net>