gsgpu: correctly set the VRAM caching attribute
I previously misunderstood what gsgpu_vram_mgr_resource::bus.caching
is used for, so the VRAM caching attributes are not set correctly.
Correct this mistake by setting VRAM caching type to write combine
if we are using the on-chip VRAM. If we are using system memory as
VRAM, then caching is set to ttm_cached.