BACKPORT: UBUNTU: SAUCE: iommu/intel: disable DMAR for SKL integrated gfx
BugLink: https://bugs.launchpad.net/bugs/2062951

This patch addresses a screen flickering issues on systems with Skylake
integrated graphics, identified under Ubuntu kernel version 6.8.0-x.
Initially thought to be a regression from kernel 6.6, it has been
determined that the flickering was caused by changes to the Ubuntu kernel
configuration, specifically the enabling of CONFIG_INTEL_IOMMU_DEFAULT_ON
and CONFIG_INTEL_IOMMU_SCALABLE_MODE_DEFAULT_ON.
77e530c1a864c ("UBUNTU: [Config] enable Intel DMA remapping by default")

The problem persists in the latest drm tip, and there is an upstream bug
addressing the issue.
https://gitlab.freedesktop.org/drm/i915/kernel/-/issues/11504

Before there is a real fix for the issue, we may have to add the affected
GPU IDs to the quirk to disable its DMAR.

Signed-off-by: Chia-Lin Kao (AceLan) <acelan.kao@canonical.com>
Acked-by: Kuan-Ying Lee <kuan-ying.lee@canonical.com>
Acked-by: Aaron Jauregui <aaron.jauregui@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>

Link: https://kernel.ubuntu.com/forgejo/ubuntu-kernel-next/linux/commit/5e37d3712ea4bb1c175177ddcb7241edd2532bdb
[Kexy: Resolved minor conflict in drivers/iommu/intel/iommu.c]
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>