AOSCOS: drm: amdgpu: use amdgpu by default for si/cik devices
On LoongArch, radeon causes a memory read violation that crashes
radeonsi_dri.so (and all processes that loaded this module). When using
amdgpu, this issue does not occur.

The upstream maintainers are likely not going to accept this change as
the default drivers were set since long ago, and we do not have enough
proof to show that this change is worth merging.

Signed-off-by: Mingcong Bai <jeffbai@aosc.io>
Link: https://t.me/c/1109254909/477558
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>