FROMLIST: cpufreq/amd-pstate: Don't update CPPC request in amd_pstate_cpu_boost_update()
When boost is changed the CPPC value is changed in amd_pstate_cpu_boost_update()
but then changed again when refresh_frequency_limits() and all it's callbacks
occur.  The first is a pointless write, so instead just update the limits for
the policy and let the policy refresh anchor everything properly.

Fixes: c8c68c38b56f ("cpufreq: amd-pstate: initialize core precision boost state")
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>

Bug: https://gitlab.freedesktop.org/drm/amd/-/issues/3690
Link: https://lore.kernel.org/linux-pm/20241012174519.897-2-mario.limonciello@amd.com/
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>