FROMEXT: arm64: dts: qcom: sc8280xp: add MDSS registers interconnect
Add the CPU-to-MDSS configuration interconnect path to the MDSS0
node on sc8280xp. This allows the display driver to explicitly
vote for the necessary bandwidth on the Config NoC to access the
MDSS registers.

Specifically, this adds the path from MASTER_APPSS_PROC to
SLAVE_DISPLAY_CFG, identified as "cpu-cfg".

Link: https://github.com/right-0903/linux-gaokun/raw/main/patch%20sets/recommended/0005-arm64-dts-qcom-sc8280xp-add-MDSS-registers-interconn.patch
Signed-off-by: Kirikaze Chiyuki <me@chyk.ink>