FROMLIST: drm: amdgpu: fix comment about amdgpu.abmlevel defaults
Since 040fdcde288a2830edc31dd507963d6aadf990d2 ("drm/amdgpu: respect the
abmlevel module parameter value if it is set"), the default value for
amdgpu.abmlevel was set to -1, or auto. However, the comment explaining
the default value was not updated to reflect the change (-1, or auto; not
-1, or disabled).

Clarify that the default value (-1) means auto.

Fixes: 040fdcde288a2830edc31dd507963d6aadf990d2 ("drm/amdgpu: respect the abmlevel module parameter value if it is set")
Reported-by: Ruikai Liu <rickliu2000@outlook.com>
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>

Link: https://lore.kernel.org/amd-gfx/43965fdb4f4a7c4223885c8e417bdc69@aosc.io/
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>