FROMGIT: drm/amdgpu: fix comment about amdgpu.abmlevel defaults
Since 040fdcde288a ("drm/amdgpu: respect the abmlevel module parameter value
if it is set"), the default value for amdgpu.abmlevel was set to -1, or auto.
However, the comment explaining the default value was not updated to reflect
the change (-1, or auto; not -1, or disabled).

Clarify that the default value (-1) means auto.

Fixes: 040fdcde288a ("drm/amdgpu: respect the abmlevel module parameter value if it is set")
Reported-by: Ruikai Liu <rickliu2000@outlook.com>
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>

(cherry picked from commit 419c896b9365be316c506335bacd6243ae42ab7b
 https://gitlab.freedesktop.org/agd5f/linux.git amd-staging-drm-next)
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>