BACKPORT: UBUNTU: SAUCE: iommu/intel: disable DMAR for KBL and CML integrated gfx
BugLink: https://bugs.launchpad.net/bugs/2086587

This patch addresses a screen flickering issues on systems with KBL and
CML integrated graphics, identified under Ubuntu kernel version 6.8.0-x.
Initially thought to be a regression from kernel 6.6, it has been
determined that the flickering was caused by changes to the Ubuntu kernel
configuration, specifically the enabling of CONFIG_INTEL_IOMMU_DEFAULT_ON
and CONFIG_INTEL_IOMMU_SCALABLE_MODE_DEFAULT_ON.
77e530c1a864c ("UBUNTU: [Config] enable Intel DMA remapping by default")

The problem persists in the latest drm tip, and there is an upstream bug
addressing the issue.
https://gitlab.freedesktop.org/drm/i915/kernel/-/issues/11504

Before there is a real fix for the issue, we may have to add the affected
GPU IDs to the quirk to disable its DMAR.

Signed-off-by: Chia-Lin Kao (AceLan) <acelan.kao@canonical.com>
Acked-by: Magali Lemes <magali.lemes@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>

Link: https://kernel.ubuntu.com/forgejo/ubuntu-kernel-next/linux/commit/aca1692216c22fa1f86e6977fee3cd6f79f730df
[Kexy: Resolved minor conflict in drivers/iommu/intel/iommu.c]
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>