AOSCOS: MIPS: lemote3a-laptop: mark as a bool and depend on DRM_RADEON
This driver, since commit "AOSCOS: MIPS: Loongson: Add a hook to turn
on/off DPMS when Fn+F7 is pressed", lemote3a-laptop uses functions from
radeon which are not exported.

Add a dependency between LEMOTE3A_LAPTOP and DRM_RADEON and mark the
former as bool so that the functions would be visible when enabled.

Fixes: "AOSCOS: MIPS: Loongson: Add a hook to turn on/off DPMS when Fn+F7 is pressed"
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>

Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>