FROMGIT: phy: phy-rockchip-samsung-hdptx: Don't use dt aliases to determine phy-id
The phy needs to know its identity in the system (phy0 or phy1 on rk3588)
for some actions and the driver currently contains code abusing of_alias
for that.

Devicetree aliases are always optional and should not be used for core
device functionality, so instead keep a list of phys on a soc in the
of_device_data and find the phy-id by comparing against the mapped
register-base.

Fixes: c4b09c562086 ("phy: phy-rockchip-samsung-hdptx: Add clock provider support")
Signed-off-by: Heiko Stuebner <heiko.stuebner@cherry.de>
Reviewed-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>
Reviewed-by: Sebastian Reichel <sebastian.reichel@collabora.com>
Link: https://lore.kernel.org/r/20241206103401.1780416-3-heiko@sntech.de
Signed-off-by: Vinod Koul <vkoul@kernel.org>

(cherry picked from commit f08d1c08563846f9be79a4859e912c8795d690fd
 https://git.kernel.org/pub/scm/linux/kernel/git/phy/linux-phy.git next)
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>