FROMLIST: irqchip/loongarch-avec:Add multi-nodes topology support
avecintc_init() enables the Advanced Interrupt Controller (AVEC) of
the boot CPU node, but nothing enables the AVEC on secondary nodes.

Move the enablement to the CPU hotplug callback so that secondary
nodes get the AVEC enabled too. In theory enabling it once per node
would be sufficient, but redundant enabling does no hard, so keep the
code simple and do it unconditionally.

Signed-off-by: Tianyang Zhang <zhangtianyang@loongson.cn>

Link: https://lore.kernel.org/all/20241230021503.14012-1-zhangtianyang@loongson.cn/
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>