AOSCOS: kvm: disable enable_virt_at_load by default
Disable enable_virt_at_load, a >= 6.12 on-by-default parameter introduced
with commit b4886fab6fb6 ("KVM: Add a module param to allow enabling
virtualization when KVM is loaded"). By enabling this paramter, VirtualBox
(as of 7.1.4) will fail to launch any virtualised systems with the
following error:

  VirtualBox can't operate in VMX root mode. Please disable the KVM kernel
  extension, recompile your kernel and reboot (VERR_VMX_IN_VMX_ROOT_MODE).

The original commit indicated that this was introduced to reduce latency
during KVM initialisation and that it "could be problematic for use case
that need to spin up VMs quickly." An optimisation that is more than
likely unnecessary for our (predominantly desktop-oriented) crowd.

Bottom line - make VirtualBox work again (until they come up with a fix).

Fixes: b4886fab6fb6 ("KVM: Add a module param to allow enabling virtualization when KVM is loaded")
Reported-by: xtex <xtexchooser@duck.com>
Link: https://github.com/AOSC-Dev/aosc-os-abbs/issues/8897
Link: https://lore.kernel.org/kvm/ZwQjUSOle6sWARsr@google.com/T/
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>

Link: https://t.me/c/1109254909/715020
Acked-by: xtex <xtexchooser@duck.com>
Reviewed-by: xtex <xtexchooser@duck.com>
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>