drm/gpuvm: Plumb through flags into drm_gpuva_init
Now that drm_gpuva_op_map has a flags field, plumb it through to the
helper that populates a drm_gpuva.

This helper is only used by drm_gpuva_remap(), which in turn is only
used by drivers which do not use flags at all (panthor, imagination),
so this change has no effect on existing drivers.

Signed-off-by: Asahi Lina <lina@asahilina.net>