LOONGSON: USB: core: Enable root_hub's remote wakeup for wakeup sources
Now we only enable the remote wakeup function for the USB wakeup source
itself at usb_port_suspend(). However on some Loongson hosts this isn't
enough to enable the S3 wakeup function for USB keyboards, so we also
enable the root_hub's remote wakeup (and disable it on error). Frankly
this is a platform-specific problem, but enable it unconditionally make
code simple and seems harmless.

Cc: stable@vger.kernel.org
Signed-off-by: Huacai Chen <chenhuacai@loongson.cn>

Link: https://github.com/chenhuacai/linux/commit/b38fbd64916064c4d4c296dafdd8ed7e76237895
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>