FROMLIST: LoongArch: Remove the deprecated notifier hook mechanism
The notifier hook mechanism in proc and cpuinfo is actually unnecessary
for LoongArch because it's not used anywhere.

It was originally added to the MIPS code in commit d6d3c9afaab4 ("MIPS:
MT: proc: Add support for printing VPE and TC ids"), and LoongArch
then inherited it.

But as the kernel code stands now, this notifier hook mechanism doesn't
really make sense for either LoongArch or MIPS.

In addition, the seq_file declaration needs to be moved to its proper
place, as only the show_ipi_list() function in smp.c requires it.

Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Tianyang Zhang <zhangtianyang@loongson.cn>
Cc: Liupu Wang <wangliupu@loongson.cn>
Cc: Bibo Mao <maobibo@loongson.cn>
Co-developed-by: Wentao Guan <guanwentao@uniontech.com>
Signed-off-by: Wentao Guan <guanwentao@uniontech.com>
Signed-off-by: Yuli Wang <wangyuli@uniontech.com>

Link: https://lore.kernel.org/all/BE6A6E80196F688F+20250209050258.2090416-3-wangyuli@uniontech.com/
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>